+++ plur1bus-hermes/src/plur1bus_hermes/__init__.py
@@ -1,6 +1,6 @@
 """Installable Hermes memory-provider plugin for PLUR1BUS."""
 
-__version__ = "7.12.7.post3"
+__version__ = "7.12.7.post4"
 
 from .provider import Plur1busMemoryProvider
 from .service import Plur1busServiceContainer